srm_backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
django-filter>=23.3
python-dotenv>=1.0
openpyxl>=3.1.0

